:truck: Move tests to new dir
