+++ packages/cli/test/integration/auth.middleware.test.ts
