+++ packages/cli/test/integration/shared/augmentation.d.ts
